moves MainController to SettingsController
